WeChat_app/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
get==2019.4.13
idna==2.8
Naked==0.1.31
post==2019.4.13
public==2019.4.13
pycryptodome==3.9.7
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
query-string==2019.4.13
requests==2.22.0
shellescape==3.8.1
sqlparse==0.3.0
urllib3==1.25.8
